OffersV2(mostly tested)
listview now handles checking if the user owns the listing when viewed. Otherwise a guest is looking at the listing and the buttons will change
accordingly.
Server code updated for checking user or guest.
Database offers design is refined.
My offers page for a user is done. Including deletion of offer, seeing
offer message, and a link to the listing page.